% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarConstructDesignMatrix.R
\name{dimarConstructDesignMatrix}
\alias{dimarConstructDesignMatrix}
\title{dimarConstructDesignMatrix}
\usage{
dimarConstructDesignMatrix(mtx)
}
\arguments{
\item{mtx}{Quantitative matrix}
}
\value{
Design matrix
}
\description{
Constructs design matrix
}
\examples{
Sample example to demonstrate the function
}
