% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarConstructReferenceData.R
\name{dimarConstructReferenceData}
\alias{dimarConstructReferenceData}
\title{dimarConstructReferenceData}
\usage{
dimarConstructReferenceData(mtx, cut = 0.2)
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{cut}{Variable cut}
}
\value{
Reference dataset
}
\description{
Constructs reference dataset
}
\examples{
Sample example to demonstrate the function
}
