% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarConstructRegularizationMatrix.R
\name{dimarConstructRegularizationMatrix}
\alias{dimarConstructRegularizationMatrix}
\title{dimarConstructRegularizationMatrix}
\usage{
dimarConstructRegularizationMatrix(design)
}
\arguments{
\item{design}{Design matrix}
}
\value{
Regularization matrix
}
\description{
Generates regularization matrix
}
\examples{
Sample example to demonstrate the function
}
