% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarEvaluatePerformance.R
\name{dimarEvaluatePerformance}
\alias{dimarEvaluatePerformance}
\title{dimarEvaluatePerformance}
\usage{
dimarEvaluatePerformance(
  Imputations,
  ref,
  sim,
  rankby = "RMSE",
  RMSEttest = TRUE,
  group = "cluster"
)
}
\arguments{
\item{Imputations}{Imputations}

\item{ref}{ref}

\item{sim}{sim}

\item{rankby}{rankby}

\item{RMSEttest}{RMSEttest}

\item{group}{group}
}
\value{
Data frame containing the following performance measures for
each imputation method: Deviation, RMSE, RSR, p-Value_F-test, Accuracy, PCC, and in case
of RMSEttest=TRUE the RMSE t-test result
}
\description{
Evaluates impuation performance.
}
\examples{
Sample example to demonstrate the function
}
