% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarGetLib.R
\name{dimarGetLib}
\alias{dimarGetLib}
\title{dimarGetLib}
\usage{
dimarGetLib(method)
}
\arguments{
\item{method}{Imputation method string}
}
\value{
R library of entered imputation method
}
\description{
Gets R library to the inputted imputation method
}
\examples{
Sample example to demonstrate the function
}
