% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarMatrixPreparation.R
\name{dimarMatrixPreparation}
\alias{dimarMatrixPreparation}
\title{dimarMatrixPreparation}
\usage{
dimarMatrixPreparation(mtx, nacut = 2, logflag = "auto")
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{nacut}{minimum number of measured data points}

\item{logflag}{should logarithm be applied (true,false,log2,log10)}
}
\description{
Prepare/Filter matrix
}
\examples{
mtx <- matrix(rnorm(1000),nrow=100)
mtx[sample(c(1:1000),100)] <- NA
mtx <- dimarMatrixPreparation(mtx,nacut=0,logflag=FALSE)

mtx <- matrix(10^(rnorm(1000,7)),nrow=100)
mtx[sample(c(1:1000),100)] <- NA
mtx <- dimarMatrixPreparation(mtx,nacut=0.8,logflag=TRUE)

}
