% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarPlotFacetHist.R
\name{dimarPlotFacetHist}
\alias{dimarPlotFacetHist}
\title{dimarPlotFacetHist}
\usage{
dimarPlotFacetHist(mtx, nbins = 20, savePlot = FALSE, width = 13, height = 5)
}
\arguments{
\item{mtx}{Numeric matrix}

\item{nbins}{Amount of intensity intervals the range of intensity values should be split into.}

\item{savePlot}{Logical. If TRUE, the histogram is saved as pdf to working directory.}

\item{width}{Width of pdf to which plot is saved if savePlot = TRUE.}

\item{height}{Height of pdf to which plot is saved if savePlot = TRUE.}
}
\value{
None
}
\description{
Plots a faceted histogram displaying the percentage of missing values for
separate intensity intervals.
For this the median protein intensity values calculated across all samples are used.
# Addapted from:
}
\examples{
library(DIMAR)
library(openxlsx)
filename <- "Test2.xlsx"
filepath <- system.file("extdata", filename, package = "DIMAR")
df <- read.xlsx(filepath, sheet="Sheet1", startRow = 2)
row.names(df) <- paste(c(1:nrow(df)), df$`Protein.(Uniprot.ID)`, sep = "_")
df <- df[, grepl("^AD\\\\d|^C\\\\d", names(df))]
mtx <- as.matrix(df)
dimarPlotFacetHist(mtx)
}
