% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarPlotHeatmap.R
\name{dimarPlotHeatmap}
\alias{dimarPlotHeatmap}
\title{dimarPlotHeatmap}
\usage{
dimarPlotHeatmap(mtx, savePlot = FALSE, width = 6, height = 6)
}
\arguments{
\item{mtx}{Numeric matrix}

\item{savePlot}{Logical. If TRUE, the histogram is saved as pdf to working directory.}

\item{width}{Width of pdf to which plot is saved if savePlot = TRUE.}

\item{height}{Height of pdf to which plot is saved if savePlot = TRUE.}
}
\value{
None
}
\description{
Plots a heatmap of the protein intensity of a dataset. The samples are displayed
in the columns and the proteins in the rows. The proteins are sorted according to an increasing
number of missing values and a decreasing mean across all samples.
}
\examples{
library(DIMAR)
library(openxlsx)
filename <- "Test2.xlsx"
filepath <- system.file("extdata", filename, package = "DIMAR")
df <- read.xlsx(filepath, sheet="Sheet1", startRow = 2)
row.names(df) <- paste(c(1:nrow(df)), df$`Protein.(Uniprot.ID)`, sep = "_")
df <- df[, grepl("^AD\\\\d|^C\\\\d", names(df))]
mtx <- as.matrix(df)
dimarPlotHeatmap(mtx)
}
