% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarPlotSampleDensity.R
\name{dimarPlotSampleDensity}
\alias{dimarPlotSampleDensity}
\title{dimarPlotSampleDensity}
\usage{
dimarPlotSampleDensity(mtx, savePlot = FALSE, width = 5, height = 13)
}
\arguments{
\item{mtx}{Numeric matrix}

\item{savePlot}{Logical. If TRUE, the histogram is saved as pdf to working directory.}

\item{width}{Width of pdf to which plot is saved if savePlot = TRUE.}

\item{height}{Height of pdf to which plot is saved if savePlot = TRUE.}
}
\value{
None
}
\description{
Plots the protein intensity distributions of each sample.
}
\examples{
library(DIMAR)
library(openxlsx)
filename <- "Test2.xlsx"
filepath <- system.file("extdata", filename, package = "DIMAR")
df <- read.xlsx(filepath, sheet="Sheet1", startRow = 2)
row.names(df) <- paste(c(1:nrow(df)), df$`Protein.(Uniprot.ID)`, sep = "_")
df <- df[, grepl("^AD\\\\d|^C\\\\d", names(df))]
mtx <- as.matrix(df)
dimarPlotSampleDensity(mtx)
}
