% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarReadInMaxQuant.R
\name{dimarReadInMaxQuant}
\alias{dimarReadInMaxQuant}
\title{dimarReadInMaxQuant}
\usage{
dimarReadInMaxQuant(filename)
}
\arguments{
\item{filename}{String of MaxQuant proteinGroups file name}
}
\value{
Quantitative matrix
}
\description{
Transforms MaxQuant proteinGroups.txt file content to quantitative matrix
}
