% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimar.R
\name{dimar}
\alias{dimar}
\title{DIMA: Data-driven selection of an imputation algorithm (Egert et al.)}
\usage{
dimar(mtx, pattern = NULL)
}
\arguments{
\item{mtx}{Data matrix or input file (.txt, .xls, .xlsx, .csv or .tsv)}

\item{pattern}{Search pattern for specifying sample names read in as default data, if not specified the user will be asked}
}
\description{
- Learn pattern of missing values
- Define reference data R with fewer MVs
- Generate patterns of missing values to R
- Apply multiple imputation algorithms
- Impute original data with best-performing imp. algorithm
}
\examples{
Imp <- dimar(mtx)
Imp <- dimar(file)
Imp <- dimar('proteinGroups.txt','LFQ')
}
