% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimar.R
\name{dimar}
\alias{dimar}
\title{dimar}
\usage{
dimar(mtx, pattern = "^Int")
}
\arguments{
\item{mtx}{Data matrix or input file (.txt, .xls, .xlsx, .csv or .tsv)}

\item{pattern}{Search pattern for specifying sample names read in as default data, if not specified the user will be asked}
}
\description{
DIMAR: Data-driven selection of an imputation algorithm in R
}
\examples{
Imp <- dimar(mtx)
Imp <- dimar(file)
Imp <- dimar('proteinGroups.txt','LFQ')
}
