% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarAssignPattern.R
\name{dimarAssignPattern}
\alias{dimarAssignPattern}
\title{dimarAssignPattern}
\usage{
dimarAssignPattern(ref, coef, mtx = NULL, npat = 0)
}
\arguments{
\item{ref}{Reference dataset}

\item{coef}{Logistic regression coefficients of missing value pattern}

\item{mtx}{Quantitative matrix}

\item{npat}{Number of patterns}
}
\value{
Reference dataset with applied pattern
}
\description{
Assigns patterns of MVs to reference data set with a logistic regression model
}
\examples{
coef <- dimarLearnPattern(mtx)
ref <- dimarConstructReferenceData(mtx)
sim <- dimarAssignPattern(ref, coef, mtx)
}
