% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarConstructDesignMatrix.R
\name{dimarConstructDesignMatrix}
\alias{dimarConstructDesignMatrix}
\title{dimarConstructDesignMatrix}
\usage{
dimarConstructDesignMatrix(mtx)
}
\arguments{
\item{mtx}{Quantitative matrix}
}
\value{
list of response vector, design matrix and vector of predictor type
}
\description{
Constructs design matrix with the factorial row/col predictors and the mean intensity
}
\examples{
design <- dimarConstructDesignMatrix(mtx)
design <- dimarConstructRegularizationMatrix(design)
fit <- glm.fit(design$X, design$y, family=binomial())

design <- dimarConstructDesignMatrix(mtx)
yhat <- exp(X$X\%\%coef)/(1+exp(X$X\%*\%coef))
}
