% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarConstructReferenceData.R
\name{dimarConstructReferenceData}
\alias{dimarConstructReferenceData}
\title{dimarConstructReferenceData}
\usage{
dimarConstructReferenceData(mtx, cut = 0.2)
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{cut}{minimum number of proteins kept for the reference data}
}
\value{
Reference dataset
}
\description{
Constructs reference dataset: Take data with least #MVs, enlarge data by proteins with least #MVs and normalize to all proteins
}
\examples{
coef <- dimarLearnPattern(mtx)
ref <- dimarConstructReferenceData(mtx)
sim <- dimarAssignPattern(ref, coef, mtx)
}
