% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarConstructRegularizationMatrix.R
\name{dimarConstructRegularizationMatrix}
\alias{dimarConstructRegularizationMatrix}
\title{dimarConstructRegularizationMatrix}
\usage{
dimarConstructRegularizationMatrix(design)
}
\arguments{
\item{design}{list of response vector, design matrix and predictor type}
}
\value{
list of response vector, design matrix with regularization and predictor type
}
\description{
Generates regularization matrix: Expand the design matrix so each predictor has at least one positive entry, this exhibits the regression coefficients to get infinite
}
\examples{
design <- dimarConstructDesignMatrix(mtx)
design <- dimarConstructRegularizationMatrix(design)
}
