% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarDoImputations.R
\name{dimarDoImputations}
\alias{dimarDoImputations}
\title{dimarDoImputations}
\usage{
dimarDoImputations(mtx, method = NULL, lib = NULL)
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{method}{Imputation method}

\item{lib}{Library}
}
\value{
Imputated dataset
}
\description{
Performs imputations
}
\examples{
Sample example to demonstrate the function
}
