% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarDoImputationsR.R
\name{dimarDoImputationsR}
\alias{dimarDoImputationsR}
\title{dimarDoImputationsR}
\usage{
dimarDoImputationsR(mtx, method = NULL, lib = NULL)
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{method}{Imputation method}

\item{lib}{R package of imputation method}
}
\value{
Quantitaive matrix, on which imputation is performed
}
\description{
Helper function for DoImputations. Actually applies the imputation functions
of various R packages to a quantitative matrix.
}
\examples{
I <- dimarDoImputationsR(mtx,'impSeq','rrcovNA')
}
