% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarDoOptimalImputation.R
\name{dimarDoOptimalImputation}
\alias{dimarDoOptimalImputation}
\title{dimarDoOptimalImputation}
\usage{
dimarDoOptimalImputation(mtx, method, lib = NULL)
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{method}{Optimal imputation method}

\item{lib}{R package of imputation method}
}
\value{
Imputed matrix
}
\description{
Performs imputation of the quantitative matrix with the given method (optimally selected by DIMA)
}
\examples{
Imputations <- dimarDoImputations(mtx)
Performance <- dimarEvaluatePerformance(Imputations, ref, mtx)
Imp <- dimarDoOptimalImputation(mtx, rownames(Performance))
}
