% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarDoOptimalImputation.R
\name{dimarDoOptimalImputation}
\alias{dimarDoOptimalImputation}
\title{dimarDoOptimalImputation}
\usage{
dimarDoOptimalImputation(mtx, method, lib = NULL)
}
\arguments{
\item{mtx}{Quantitative matrix}

\item{method}{Imputation method}

\item{lib}{Library}
}
\value{
Quantitative matrix on which optimal imputation method was applied.
}
\description{
Performs the optimal imputation to quantitative matrix.
}
\examples{
Sample example to demonstrate the function
}
