% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarLearnPattern.R
\name{dimarLearnPattern}
\alias{dimarLearnPattern}
\title{dimarLearnPattern}
\usage{
dimarLearnPattern(mtx)
}
\arguments{
\item{mtx}{Quantitative matrix}
}
\value{
Logistic regression coefficients describing missing value pattern
}
\description{
Learns missing value occurence pattern from quantitative matrix by a logistic regression model
}
\examples{
#' mtx <- matrix(rnorm(1000),nrow=100)
mtx[sample(c(1:1000),100)] <- NA
coef <- dimarLearnPattern(mtx)
}
