% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dimarLearnPattern.R
\name{dimarLearnPattern}
\alias{dimarLearnPattern}
\title{dimarLearnPattern}
\usage{
dimarLearnPattern(mtx)
}
\arguments{
\item{mtx}{Quantitative matrix}
}
\value{
Logistic regression coefficients describing missing value pattern
}
\description{
Learns missing value pattern from quantitative matrix-
}
\examples{
Sample example to demonstrate the function
}
